Remove old texts
They are incorrect since used tecnique forgets about filtration on homology, I just computed homology of maximal complex in filtration again.

But it can be fixed via the following scheme of work:
1. Recall results on convergence of Mayer-Vietoris spectral sequence (which are prior to similar results of Alvaro Torres eight years after).
2. Compute page on which our \varepsilon acyclic cover by two open sets will converge.
3. Apply the result.

In sense of the paper it means recollection of data about spectral sequence and error propagation in it in a separate section of preliminaries.
After that scheme is as it used to be: recollect proofs by Barmak, in detail write them in both cases. And then in detail write what is different in our case.
Problem is solvable, scheme of the proof is still correct.

Yet another nuance: it seems essential that computations are over a field. It can be avoided by considering a s.e.s Z -> Q -> Z(\infty).
I tend to believe there is no distinguished trivial approximate morphism in case of arbitrary ring since we do not have classification of graded ideals.