Write appendix on error-propagation in Mayer-Vietoris spectral sequence
Seems that I was wrongly blaming myself in older commits --- where is no problem in using Mayer-Vietoris l.e.s since it is used only to establish induced equivalence and reasoning about error in l.e.s. of graded modules with graded maps is valid.
There still may be new problems while proving existence of the map in reverse direction but even formula for Tor-functors was already covered.

However old proof is still non-satisfactory because notion of a mapping cylinder of a poset map should be replaced with its persistence analogue.
This change will yield conditions on posets. Issue --- relax these conditions.

Also it is now again clear that current proof has no place for spectral sequences. This is why it is after results now.

So scheme from this point:
1. Translate and revise proof of approximate Lemma 3.
2. Introduce persistense mapping cylinder of map of persistense posets (in easiest fashion it's a simply cylinder of a filtered map).
3. Write the proof for fields --- it will only a little differ from old proofs.
4. Write proof for Z via main theorem and s.e.s Z->Q->Z_{inf}. It will probably be separate theorem since there may be another error multiple.